Changed layout of files
