+++ src/index.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import App from './App';
-import reportWebVitals from './reportWebVitals';
+import App from './SolarSystem/App';
+import reportWebVitals from './Misc/reportWebVitals';
 
 ReactDOM.render(
   <React.StrictMode>
